Policies a meio

+++ server-api/routes/api.php
@@ -27,37 +27,36 @@ Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
 });
 
 Route::post('login', [AuthController::class, 'login']);
-Route::post('vcards', [VcardController::class, 'store']);
+Route::post('vcards', [VcardController::class, 'store'])->middleware('can:create,App\Models\Vcard');
 Route::middleware('auth:api')->group(function () {
     
     Route::post('logout', [AuthController::class, 'logout']);
     Route::get('users/me', [UserController::class, 'show_me']);
 
     //Admins
-    Route::post('admins', [AdminController::class,'store']);
-    Route::put('admins/{admin}', [AdminController::class,'update']);
-    Route::delete('admins/{admin}', [AdminController::class,'destroy']);
-    Route::patch('admins/{admin}/password', [AdminController::class,'update_password']);
+    Route::post('admins', [AdminController::class,'store'])->middleware('can:create,App\Models\Admin');
+    Route::put('admins/{admin}', [AdminController::class,'update'])->middleware('can:update,admin');
+    Route::delete('admins/{admin}', [AdminController::class,'destroy'])->middleware('can:delete,admin');
+    Route::patch('admins/{admin}/password', [AdminController::class,'update_password'])->middleware('can:updatePassword,admin');
 
     //Users
     Route::get('users', [UserController::class, 'index'])->middleware('can:viewAny,App\Models\User');
     Route::get('users/{user}', [UserController::class, 'show'])->middleware('can:view,user');
-    Route::get('distributionOfUsers', [UserController::class, 'getDistributionOfUsers']);
-    // Route::put('users/{user}', [UserController::class, 'update'])->middleware('can:update,user');
+    Route::get('distributionOfUsers', [UserController::class, 'getDistributionOfUsers'])->middleware('can:getDistributionOfUsers,App\Models\User');
     
     //Vcards
-    Route::get('vcards', [VcardController::class, 'index']);
-    Route::get('vcards/{vcard}', [VcardController::class, 'show']);
-    Route::put('vcards/{vcard}', [VcardController::class, 'update']);
-    Route::patch('vcards/{vcard}/maxDebit', [VcardController::class, 'updateMaxDebit']);
-    Route::patch('vcards/{vcard}/blocked', [VCardController::class, 'updateBlocked']);
-    Route::patch('vcards/{vcard}/profile', [VCardController::class, 'updateProfile']);
-    Route::get('vcards/{vcard}/transactions', [VCardController::class, 'getTransactionsOfVcard']);
-    Route::delete('vcards/{vcard}', [VCardController::class,'destroy']);
-    Route::patch('vcards/{vcard}/password', [VCardController::class, 'update_password']);
-    Route::patch('vcards/{vcard}/confirmationCode', [VCardController::class, 'update_confirmation_code']);
-    Route::get('vcards/{vcard}/categories', [VCardController::class, 'getCategoryOfVcard']);
-    Route::get('vcardsActive', [VCardController::class, 'getActiveVcards']);
+    Route::get('vcards', [VcardController::class, 'index'])->middleware('can:viewAny,App\Models\Vcard');
+    Route::get('vcards/{vcard}', [VcardController::class, 'show'])->middleware('can:view,vcard');
+    Route::put('vcards/{vcard}', [VcardController::class, 'update'])->middleware('can:update,vcard');
+    // Route::patch('vcards/{vcard}/maxDebit', [VcardController::class, 'updateMaxDebit']);
+    Route::patch('vcards/{vcard}/blocked', [VCardController::class, 'updateBlocked'])->middleware('can:updateBlocked,vcard');
+    Route::patch('vcards/{vcard}/profile', [VCardController::class, 'updateProfile'])->middleware('can:updateProfile,vcard');
+    Route::get('vcards/{vcard}/transactions', [VCardController::class, 'getTransactionsOfVcard'])->middleware('can:getTransactions,vcard');
+    Route::delete('vcards/{vcard}', [VCardController::class,'destroy'])->middleware('can:delete,vcard');
+    Route::patch('vcards/{vcard}/password', [VCardController::class, 'update_password'])->middleware('can:updatePassword,vcard');
+    Route::patch('vcards/{vcard}/confirmationCode', [VCardController::class, 'update_confirmation_code'])->middleware('can:updateConfirmationCode,App\Models\Vcard');
+    Route::get('vcards/{vcard}/categories', [VCardController::class, 'getCategoryOfVcard'])->middleware('can:getCategories,vcard');
+    Route::get('vcardsActive', [VCardController::class, 'getActiveVcards'])->middleware('can:getActiveVcards,App\Models\Vcard');;
 
     //Transactions
 
@@ -80,8 +79,8 @@ Route::middleware('auth:api')->group(function () {
     Route::put('defaultCategories/{defaultCategory}', [DefaultCategoryController::class, 'update']);
     Route::delete('defaultCategories/{defaultCategory}', [DefaultCategoryController::class, 'delete']);
     
-    Route::get('categories/{category}', [CategoryController::class, 'show']);
     Route::get('categories', [CategoryController::class, 'index']);
+    Route::get('categories/{category}', [CategoryController::class, 'show']);
     Route::post('categories', [CategoryController::class, 'store']);
     Route::put('categories/{category}', [CategoryController::class, 'update']);
     Route::delete('categories/{category}', [CategoryController::class, 'delete']);
